pull update

+++ api/cas-server-core-api-configuration-model/src/main/java/org/apereo/cas/configuration/model/support/consent/ConsentProperties.java
@@ -6,11 +6,13 @@ import org.apereo.cas.configuration.model.support.ldap.AbstractLdapSearchPropert
 import org.apereo.cas.configuration.model.support.mongo.SingleCollectionMongoDbProperties;
 import org.apereo.cas.configuration.support.RequiresModule;
 import org.apereo.cas.configuration.support.SpringResourceProperties;
-import org.springframework.boot.context.properties.NestedConfigurationProperty;
-import java.io.Serializable;
-import java.time.temporal.ChronoUnit;
+
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.boot.context.properties.NestedConfigurationProperty;
+
+import java.io.Serializable;
+import java.time.temporal.ChronoUnit;
 
 /**
  * This is {@link ConsentProperties}.
@@ -29,7 +31,7 @@ public class ConsentProperties implements Serializable {
      * Global reminder time unit, to reconfirm consent
      * in cases no changes are detected.
      */
-    private int reminder = 30;
+    private long reminder = 30;
 
     /**
      * Global reminder time unit of measure, to reconfirm consent
